requirements.txt
absl-py==2.1.0
accelerate==1.0.1
adtk==0.6.2
aenum==3.1.15
agentql==1.6.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anthropic==0.37.1
anyio==4.6.2.post1
arxiv==2.1.3
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
av==12.3.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.5.0
category-encoders==2.6.4
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
ctranslate2==4.4.0
cycler==0.12.1
Cython==3.0.11
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.7
datasets==3.0.1
decorator==5.1.1
deepgram-sdk==3.7.5
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==6.3.5
evaluate==0.4.3
executing==2.1.0
faiss-cpu==1.9.0.post1
fake-http-header==0.3.5
faster-whisper==1.0.3
fastjsonschema==2.20.0
feedparser==6.0.11
filelock==3.16.1
firecrawl-py==1.5.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.1
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
imbalanced-learn==0.12.4
importlib_metadata==8.5.0
iniconfig==2.0.0
instructor==1.6.4
ipython==8.29.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.6.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kaleido==0.2.1
keras==3.7.0
kiwisolver==1.4.7
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.25
langchain-ollama==0.2.1
langchain-openai==0.2.9
langchain-text-splitters==0.3.2
langgraph==0.2.52
langgraph-checkpoint==2.0.5
langgraph-sdk==0.1.36
langsmith==0.1.143
libclang==18.1.1
lightgbm==4.5.0
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mockito==1.5.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.18.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
ollama==0.4.4
ollama-ocr==0.1.3
onnxruntime==1.19.2
openai==1.54.4
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.10
packaging==24.1
pandas==2.1.4
parso==0.8.4
patsy==0.5.6
pdf2image==1.17.0
pillow==11.0.0
platformdirs==4.3.6
playwright==1.48.0
plotly==5.24.1
plotly-resampler==0.10.0
pluggy==1.5.0
pmdarima==2.0.4
primp==0.7.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.2
psutil==6.1.0
pure_eval==0.2.3
pyarrow==17.0.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
pyee==12.0.0
Pygments==2.18.0
PyMuPDF==1.24.14
pyod==2.0.2
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
pyreadline3==3.5.4
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mockito==0.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-magic==0.4.27
pytz==2024.2
pywin32==308
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rich==13.9.4
rpds-py==0.20.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.13.2
sendgrid==6.11.0
sentencepiece==0.2.0
setuptools==75.2.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
sktime==0.26.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
speechbrain==1.0.1
SpeechRecognition==3.11.0
SQLAlchemy==2.0.35
srt==3.5.3
stack-data==0.6.3
starkbank-ecdsa==2.2.0
statsmodels==0.14.4
streamlit==1.41.1
sympy==1.13.1
tabulate==0.9.0
tbats==1.1.3
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-cpu==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf-playwright-stealth==1.1.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tiktoken==0.8.0
timedelta==2020.12.3
tokenizers==0.20.1
toml==0.10.2
torch==2.5.0
torchaudio==2.5.0
tornado==6.4.2
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
tsdownsample==0.1.3
typer==0.13.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vosk==0.3.45
watchdog==6.0.0
wcwidth==0.2.13
websockets==13.1
Werkzeug==3.0.6
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.0
xxhash==3.5.0
yarl==1.15.5
yellowbrick==1.5
zipp==3.20.2
